Added errors, refactoring.
