--- parsco/src/internal/mod.rs
@@ -1,6 +1,0 @@
-//! This module is purely for internal organization purposes.
-
-pub mod basic;
-pub mod control;
-pub mod delimited;
-pub mod repeating;
