+++ parsco/src/lib.rs
@@ -73,24 +73,14 @@ use std::ops::Range;
 
 use common::Void;
 
-pub use internal::basic::{tag, sym, fst, dbg, fun, constant};
-pub use internal::delimited::{preceded, terminated, delimited};
-pub use internal::repeating::{many0, many1, list0, take_while0, take_while1, take_until, ws, take, take_nm};
-pub use internal::control::{alt, map, flat_map, eat, opt, satisfying, flat_map_err};
-
-/// The structs used for the actual parsing.
-/// 
-/// User of the crate shouldn't have need to use these directly, but use functions that create them instead.
-pub mod parsers {
-    pub use internal::control::{Alt, Empty, Map, FlatMap, Eat, Opt, Satisfying, FlatMapErr};
-    pub use internal::delimited::{Preceded, Terminated, Delimited};
-    pub use internal::repeating::{Many0, Many1, List0, TakeWhile0, TakeWhile1, TakeUntil, Whitespace, Take, TakeNM};
-    pub use internal::basic::{Tag, Symbol, Fst, Fun, Dbg, Constant};
-}
+pub use parsers::basic::{tag, sym, fst, dbg, fun, constant};
+pub use parsers::delimited::{preceded, terminated, delimited};
+pub use parsers::repeating::{many0, many1, list0, take_while0, take_while1, take_until, ws, take, take_nm};
+pub use parsers::control::{alt, map, flat_map, eat, opt, satisfying, flat_map_err};
 
 pub mod common;
 
-mod internal;
+pub mod parsers;
 
 type Place = usize;
