+++ parsco/src/parsers/mod.rs
@@ -1,0 +1,13 @@
+//! The structs used for the actual parsing.
+//! 
+//! User of the crate shouldn't have need to use these directly, but use functions that create them instead.
+
+pub(crate) mod basic;
+pub(crate) mod control;
+pub(crate) mod delimited;
+pub(crate) mod repeating;
+
+pub use self::control::{Alt, Empty, Map, FlatMap, Eat, Opt, Satisfying, FlatMapErr};
+pub use self::delimited::{Preceded, Terminated, Delimited};
+pub use self::repeating::{Many0, Many1, List0, TakeWhile0, TakeWhile1, TakeUntil, Whitespace, Take, TakeNM};
+pub use self::basic::{Tag, Symbol, Fst, Fun, Dbg, Constant};
